Refactor staff controllers/tests to agent
